productList servlet update(change name, location),findByName modificated, class names fixed
