+++ webapp/src/main/java/com/infoshare/fourfan/servlet/AdminEditProductServlet2.java
@@ -21,7 +21,7 @@ import java.util.Map;
 import java.util.logging.Logger;
 
 @WebServlet("/admin-edit-product")
-public class adminEditProductServlet2 extends HttpServlet {
+public class AdminEditProductServlet2 extends HttpServlet {
 
     @Inject
     ProductService productService;
@@ -29,7 +29,7 @@ public class adminEditProductServlet2 extends HttpServlet {
     @Inject
     private TemplateProvider templateProvider;
 
-    private static final Logger logger = Logger.getLogger(adminEditProductServlet2.class.getName());
+    private static final Logger logger = Logger.getLogger(AdminEditProductServlet2.class.getName());
 
 
     @Override
