+++ webapp/src/main/java/com/infoshare/fourfan/servlet/ProductListServlet.java
@@ -1,0 +1,46 @@
+package com.infoshare.fourfan.servlet;
+
+import com.infoshare.fourfan.freemarker.TemplateProvider;
+import com.infoshare.fourfan.service.ProductService;
+import freemarker.template.Template;
+import freemarker.template.TemplateException;
+
+import javax.inject.Inject;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.logging.Logger;
+
+@WebServlet("/productList")
+public class ProductListServlet extends HttpServlet {
+
+    @Inject
+    private ProductService productService;
+
+    @Inject
+    private TemplateProvider templateProvider;
+
+    private static final Logger logger = Logger.getLogger(ProductListServlet.class.getName());
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        resp.setContentType("text/html;charset=UTF-8");
+
+        Template template = templateProvider.getTemplate(getServletContext(), "productList.ftlh");
+        PrintWriter printWriter = resp.getWriter();
+
+        Map<String, Object> dataModel = new HashMap<>();
+        dataModel.put("products", productService.findAllJson());
+
+        try {
+            template.process(dataModel, printWriter);
+        } catch (TemplateException e) {
+            logger.severe(e.getMessage());
+        }
+    }
+}
